Fix .scss files to work with Sass 3.4.5.
Two bugs here, relative to Sass 3.4.5 (Selective Steve).
The first bug, reported at
https://github.com/dsociative/2048/commit/80258ad81716fd0dc263a3b0e154c7624afe1a0e
is that Sass apparently requires interpolation braces #{} around
a mixin parameter used on the left side of a colon.

The second bug, visible at that same commit, is that when you
refer to an unbound variable inside a mixin, Sass looks it up
in the global scope rather than in the local scope of the "caller"
of the mixin. For example,

    @mixin foo() { bar: $bar; }

    $bar: 1;
    a {
      $bar: 2;
      @include foo();
      baz: $bar;
    }

gives `a { bar: 1; baz: 2; }`.

According to the Sass maintainers, this is by design. The fix is to
explicitly mutate that global variable via `!global` rather than
trying to get the mixin to use a local variable.
https://github.com/sass/sass/issues/1474